application_switch: Provide access to F1–F4 keys
Since I overrode the F1 through F4 keys globally, there was no way to
access press a non-custom variant of these keys. Realised this when
trying to access the manual in Mutt, which is bound to F1.

Add Shift-F1–F4 bindings that map to F1–F4, in case we need access to
the original set of keys.

Needed to add a `doAfter` delay after disabling the binding because the
delay seems to take a bit to go through. Thanks to 'knu'
(https://github.com/knu) for the tip:
https://github.com/Hammerspoon/hammerspoon/issues/1252#issuecomment-290411701

Also gleaned it from:
https://stackoverflow.com/questions/40986242/key-repeats-are-delayed-in-my-hammerspoon-script